Fixed a bug with specular sampling also added a few more hdr backgrounds but they dont work smile

+++ Core/FrameBuffer.cpp
@@ -84,7 +84,7 @@ namespace Gep
     [[nodiscard]] FrameBuffer FrameBuffer::CreateScreenHDR(const glm::ivec2 size)
     {
         FrameBuffer fb = Create(size);
-        fb.AddTexture(GL_COLOR_ATTACHMENT0, GL_RGBA16F, GL_RGBA, GL_FLOAT);
+        fb.AddTexture(GL_COLOR_ATTACHMENT0, GL_RGBA32F, GL_RGBA, GL_FLOAT);
         fb.AddTexture(GL_DEPTH_ATTACHMENT, GL_DEPTH_COMPONENT, GL_DEPTH_COMPONENT, GL_FLOAT);
         return fb;
     }
